wip: ws sever 3
