+++ commm/client.go
@@ -1,12 +1,7 @@
 package main
 
-import "github.com/gorilla/websocket"
+import (
+	"github.com/gorilla/websocket"
+	"time"
+)
 
-type Client struct {
-	// hub
-	hub *Hub
-	// WS conn
-	conn *websocket.Conn
-	// Buffered channel of outbound messages
-	send chan []byte
-}
